requirements/base.txt
adrf==0.1.5
amqp==5.2.0
anyio==4.2.0
asgiref==3.7.2
async-property==0.2.2
async-timeout==4.0.3
billiard==4.2.0
celery==5.3.6
certifi==2023.11.17
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
coreapi==2.3.3
coreschema==0.0.4
cron-descriptor==1.4.3
cryptography==42.0.2
defusedxml==0.7.1
Deprecated==1.2.14
dj-database-url==2.1.0
dj-rest-auth==5.0.2
Django==4.2.10
django-allauth==0.57.0
django-celery-beat==2.5.0
django-cors-headers==4.3.1
django-elasticsearch-dsl==8.0
django-json-widget==1.1.1
django-redis==5.4.0
django-rest-framework==0.1.0
django-rest-swagger==2.2.0
django-timezone-field==6.1.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
dnspython==2.4.2
drf-yasg==1.20.0
elastic-transport==8.12.0
elasticsearch==8.12.1
elasticsearch-dsl==8.12.0
exceptiongroup==1.2.0
future==1.0.0
h11==0.14.0
idna==3.6
inflection==0.5.1
itypes==1.2.0
Jinja2==3.1.2
jwcrypto==1.5.1
kombu==5.3.4
Markdown==3.4.3
MarkupSafe==2.1.3
oauthlib==3.2.2
openapi-codec==1.3.2
packaging==23.2
pillow==10.2.0
pipdeptree==2.13.1
prompt-toolkit==3.0.43
psycopg2-binary==2.9.9
pycparser==2.21
PyJWT==2.8.0
python-crontab==3.0.0
python-dateutil==2.8.2
python-dotenv==1.0.0
python-http-client==3.3.7
python3-openid==3.2.0
pytz==2023.3.post1
redis==5.0.1
requests==2.31.0
requests-oauthlib==1.3.1
ruamel.yaml==0.17.31
ruamel.yaml.clib==0.2.7
simplejson==3.19.1
six==1.16.0
sniffio==1.3.0
sqlparse==0.4.4
starkbank-ecdsa==2.2.0
typing_extensions==4.9.0
tzdata==2023.4
uritemplate==4.1.1
urllib3==2.1.0
uvicorn==0.27.0.post1
vine==5.1.0
watchfiles==0.21.0
wcwidth==0.2.12
whitenoise==6.6.0
wrapt==1.16.0
